+++ tests/test_utils.py
@@ -1,4 +1,5 @@
 import pytest
+import re
 import sys
 
 from unittest.mock import patch
@@ -71,13 +72,12 @@ def test_validate_unit_group_invalid(data_store):
 
 # Tests 'get_converter_units' function
 def test_get_converter_units_valid(data_store, conversion_data):
-    with patch("unit_converter.utils.get_users_input", side_effect=["meters", "miles"]):
-        result = get_converter_units(data_store, conversion_data)
-        assert result == "meters", "miles"
+    with patch("unit_converter.utils.get_users_input", side_effect=["meters", "miles"]):        
+        assert get_converter_units(data_store, conversion_data) == ("meters", "miles")
 
 def test_get_converter_units_empty(data_store, conversion_data):
     with patch("unit_converter.utils.get_users_input", side_effect=["", "miles"]):
-        with pytest.raises(ValueError, match="Unit type cannot be empty!"):
+        with pytest.raises(ValueError, match="'unit_type' cannot be empty!"):
             get_converter_units(data_store, conversion_data)
 
 def test_get_converter_units_invalid(data_store, conversion_data):
@@ -91,44 +91,44 @@ def test_get_amount_integer(conversion_data):
     with patch("unit_converter.utils.get_users_input", return_value="10"):
         result = get_amount(conversion_data)
         assert result == 10.0
-        assert conversion_data.amount == "10"
+        assert conversion_data.amount == 10.0
 
 def test_get_amount_decimal(conversion_data):
     with patch("unit_converter.utils.get_users_input", return_value="10.0"):
         result = get_amount(conversion_data)
         assert result == 10.0
-        assert conversion_data.amount == "10.0"
+        assert conversion_data.amount == 10.0
 
 def test_get_amount_negative(conversion_data):
     with patch("unit_converter.utils.get_users_input", return_value="-10"):
         result = get_amount(conversion_data)
         assert result == -10.0
-        assert conversion_data.amount == "-10"
+        assert conversion_data.amount == -10.0
 
 def test_get_amount_empty(conversion_data):
     with patch("unit_converter.utils.get_users_input", return_value=""):
-        with pytest.raises(ValueError, match="Invalid amount! Please, insert integer or decimals! (e.g. 10 or 10.0)"):
+        with pytest.raises(ValueError, match=re.escape("Invalid amount! Please, insert integer or decimals! (e.g. 10 or 10.0)")):
             get_amount(conversion_data)
 
 def test_get_amount_invalid_format(conversion_data):
     with patch("unit_converter.utils.get_users_input", return_value="10."):
-        with pytest.raises(ValueError, match="Invalid amount! Please, insert integer or decimals! (e.g. 10 or 10.0)"):
+        with pytest.raises(ValueError, match=re.escape("Invalid amount! Please, insert integer or decimals! (e.g. 10 or 10.0)")):
             get_amount(conversion_data)
 
 def test_get_amount_str(conversion_data):
     with patch("unit_converter.utils.get_users_input", return_value="ten"):
-        with pytest.raises(ValueError, match="Invalid amount! Please, insert integer or decimals! (e.g. 10 or 10.0)"):
+        with pytest.raises(ValueError, match=re.escape("Invalid amount! Please, insert integer or decimals! (e.g. 10 or 10.0)")):
             get_amount(conversion_data)
 
 
 # Test 'resolve_aliases' function
-def test_resolve_aliases_valid_unit():
+def test_resolve_aliases_valid_unit(data_store):
     assert resolve_aliases(data_store, "length", "meters") == "meters"
 
-def test_resolve_aliases_valid_alias():
+def test_resolve_aliases_valid_alias(data_store):
     assert resolve_aliases(data_store, "length", "m") == "meters"
 
-def test_resolve_aliases_invalid():
+def test_resolve_aliases_invalid(data_store):
     assert resolve_aliases(data_store, "length", "invalid") is False
 
 
@@ -137,7 +137,7 @@ def test_parse_time_input_complete():
     assert parse_time_input("10h:10m:10s") == 10 * 3600 + 10 * 60 + 10
 
 def test_parse_time_input_no_hours():
-    assert parse_time_input("10m:10s") == 10 * 60 + 10
+    assert parse_time_input(":10m:10s") == 10 * 60 + 10
 
 def test_parse_time_input_complete_no_minutes():
     assert parse_time_input("10h:10s") == 10 * 3600 + 10
@@ -149,10 +149,10 @@ def test_parse_time_input_complete_only_hours():
     assert parse_time_input("10h") == 10 * 3600
 
 def test_parse_time_input_complete_only_minutes():
-    assert parse_time_input("10m") == 10 * 60
+    assert parse_time_input(":10m") == 10 * 60
 
 def test_parse_time_input_complete_only_seconds():
-    assert parse_time_input("10s") == 10
+    assert parse_time_input(":10s") == 10
 
 def test_parse_time_input_invalid_format():
     assert parse_time_input("10h-10m-10s") is None
@@ -164,7 +164,7 @@ def test_parse_time_input_str():
     assert parse_time_input("Ten Hours") is None
 
 def test_parse_time_input_empty():
-    assert parse_time_input("") is None
+    assert parse_time_input(" ") is None
 
 
 # Test 'check_time_is_none' function
@@ -178,7 +178,7 @@ def test_check_time_is_none_none():
 
 # Test 'get_seconds' function
 def test_get_seconds_valid(data_store):
-    assert get_seconds(data_store, "time", 10, 10, 10) == 10 * 365.2425 + 10 * 30.436875 + 10 * data_store.units["time"]["days"]
+    assert get_seconds(data_store, "time", 10, 10, 10) == (10 * 365.2425 * 86400.0) + (10 * 30.436875 * 86400.0) + (10 * 86400.0)
 
 def test_get_seconds_all_zeroes(data_store):
     assert get_seconds(data_store, "time", 0, 0, 0) == 0
@@ -304,6 +304,3 @@ def test_get_index_from_month_invalid(data_store):
 # Test 'gets_days_from_index' function
 def test_get_days_from_index_valid(data_store):
     assert gets_days_from_index(data_store, "12") == 31
-
-def test_get_days_from_index_invalid(data_store):
-    assert gets_days_from_index(data_store, "13") is None